Extrat methods responsible for creating charts to additional instance and cleaned run file (moved functions to utils)
