+++ app/run.py
@@ -1,110 +1,15 @@
-import logging
 import requests
 
 from api_manager import ApiManager
 from file_manager import FileManager
-from utils import get_logger, parse_args, read_config
-
-REQUIRED_LAUREATES_DATA = {
-    "givenName": "en",
-    "familyName": "en",
-    "gender": None,
-    "birth": "date",
-    "wikipedia": "english",
-    "nobelPrizes": {"awardYear", "category", "prizeStatus", "motivation"}
-}
+from utils import get_logger, parse_args, read_config, get_laureates_necessary_data, fetch_nobel_laureates
 
 
-def get_laureates_necessary_data(logger: logging.Logger, laureates_data: list[dict]) -> list[dict]:
+def start_script() -> None:
     """
-    Extracts necessary data from the fetched Nobel laureates data. If any of the required data is missing for a given
-    laureate data, it is added as 'Unknown'.
-
-    :param logger: A logger instance for printing output messages.
-    :param laureates_data: Laureates data fetched from API.
-    :return: List of dictionaries containing trimmed laureates data only with necessary fields.
+    Reads run script command arguments, creates logger, reads config and creates Api and File Managers instances.
+    Fetch data from the API and writes it to .json and .excel ( if special flags were set in the script run command ).
     """
-    new_laureates_data = []
-    for laureate in laureates_data:
-        # Skip organizations
-        if "orgName" in laureate:
-            logger.debug(f"Skipping organization data: {laureate['orgName']['en']}")
-            continue
-
-        necessary_laureates_data = {}
-        for key, value in REQUIRED_LAUREATES_DATA.items():
-            if key in laureate:
-                if key == "nobelPrizes":
-                    necessary_laureates_data[key] = trim_nobel_prizes_data(laureate['nobelPrizes'])
-                else:
-                    if isinstance(laureate[key], dict) and value in laureate[key]:
-                        necessary_laureates_data[key] = laureate[key][value]
-                    else:
-                        necessary_laureates_data[key] = laureate[key]
-            else:
-                necessary_laureates_data[key] = "Unknown"
-
-        new_laureates_data.append(necessary_laureates_data)
-
-    return new_laureates_data
-
-
-def trim_nobel_prizes_data(laureate_prizes: list[dict]) -> list[dict]:
-    """
-    Trims laureate prizes data to include only required fields.
-
-    :param laureate_prizes: List of dictionaries containing laureate prizes data.
-    :return: Trimmed laureate prizes data.
-    """
-    trimmed_nobel_prizes_data = []
-    for nobel_prize in laureate_prizes:
-        trimmed_nobel_prize_data = {}
-        for prize_data in REQUIRED_LAUREATES_DATA['nobelPrizes']:
-            if prize_data in nobel_prize:
-                if isinstance(nobel_prize[prize_data], dict) and "en" in nobel_prize[prize_data]:
-                    trimmed_nobel_prize_data[prize_data] = nobel_prize[prize_data]['en']
-                else:
-                    trimmed_nobel_prize_data[prize_data] = nobel_prize[prize_data]
-        trimmed_nobel_prizes_data.append(trimmed_nobel_prize_data)
-
-    return trimmed_nobel_prizes_data
-
-
-def fetch_nobel_laureates(logger: logging.Logger, api_mgr: ApiManager, url_api_params: dict,
-                          max_api_attempts: int) -> list:
-    """
-    Fetches Nobel laureates data from the specified API and given parameters.
-
-    :param logger: A logger instance for printing output messages.
-    :param api_mgr: Instance of the ApiManager class responsible for making API requests.
-    :param url_api_params: Url parameters for the API request.
-    :param max_api_attempts: Max API request attempts
-    :return: List of dictionaries containing data about Nobel laureates.
-    """
-    fetched_data = []
-    attempts = 0
-    logger.info(f"Fetching nobel laureates from year {url_api_params['nobelPrizeYear']} to {url_api_params['yearTo']}")
-    while attempts < max_api_attempts:
-        attempts += 1
-        logger.debug(f"Attempt {attempts}/{max_api_attempts}:")
-        try:
-            laureates_data = api_mgr.get_laureates_data(url_api_params)
-            if not laureates_data:
-                logger.debug("All data for the set parameters has been fetched. Skipping other attempts.")
-                break
-
-            fetched_data.extend(laureates_data)
-            url_api_params['offset'] += 50
-        except requests.exceptions.RequestException as e:
-            logger.error(f"An error occurred during attempt {attempts} to fetch data from API: {e}")
-            if attempts == max_api_attempts:
-                logger.error(f"Max number of attempts ({max_api_attempts}) reached. Cannot fetch data from API.")
-                break
-
-    return fetched_data
-
-
-if __name__ == '__main__':
     args = parse_args()
     logger = get_logger("FETCHER", min(args.v * 10, 50))
     cfg = read_config("config.toml")
@@ -131,3 +36,7 @@ if __name__ == '__main__':
         logger.error("No data was retrieved from the API!")
 
     logger.info("Fetcher finished his work.")
+
+
+if __name__ == '__main__':
+    start_script()
